Tweak the `requireAuthentication` check in the application route
It causes issues in Ember v4 because it's the first route that's entered (and the login route is nested under it). Ideally this check happens in each route, but that's not possible for the engine route. To work around the issue we check the `transition.to` to see if the target is the engine route, and only then do the redirect if needed.