Add support for "English accounts model"
Not all projects use the Dutch version of the "Accounts" model, so with the introduction of the impersonation feature we prevented some apps from updating to the latest dashboard.

We now support both English and Dutch versions, but stick to Dutch by default for now.

The english version can be enabled by replacing the `{{EMBER_ENGLISH_ACCOUNTS_MODEL}}` to `'true'`.